final commit

+++ industry-service/src/main/java/com/petr/industryservice/dto/IndustryDto.java
@@ -1,5 +1,5 @@
 package com.petr.industryservice.dto;
 
-public record IndustryDto(int id,
+public record IndustryDto(Long id,
                           String name) {
 }
